feat(HW-882): Tokens re-design
