+++ src/components/token/token.types.ts
@@ -1,0 +1,11 @@
+import {IToken} from '@app/types';
+
+export type TokenRowProps = {
+  item: IToken;
+  checked?: boolean;
+  onPress?: () => void;
+};
+
+export type TokenIconProps = {
+  item: IToken;
+};
